doma/src/test/java/org/seasar/doma/internal/jdbc/command/EntityResultListHandlerTest.java
package org.seasar.doma.internal.jdbc.command;

import java.util.List;

import org.seasar.doma.domain.IntegerDomain;
import org.seasar.doma.domain.StringDomain;
import org.seasar.doma.internal.jdbc.command.EntityResultListHandler;
import org.seasar.doma.internal.jdbc.mock.ColumnMetaData;
import org.seasar.doma.internal.jdbc.mock.MockConfig;
import org.seasar.doma.internal.jdbc.mock.MockResultSet;
import org.seasar.doma.internal.jdbc.mock.MockResultSetMetaData;
import org.seasar.doma.internal.jdbc.mock.RowData;
import org.seasar.doma.internal.jdbc.query.SqlFileSelectQuery;
import org.seasar.doma.internal.jdbc.sql.SqlFiles;

import junit.framework.TestCase;
import example.entity.Emp;
import example.entity.Emp_;

/**
 * @author taedium
 * 
 */
public class EntityResultListHandlerTest extends TestCase {

    private MockConfig runtimeConfig = new MockConfig();

    public void testHandle() throws Exception {
        MockResultSetMetaData metaData = new MockResultSetMetaData();
        metaData.columns.add(new ColumnMetaData("id"));
        metaData.columns.add(new ColumnMetaData("name"));
        MockResultSet resultSet = new MockResultSet(metaData);
        resultSet.rows.add(new RowData(1, "aaa"));
        resultSet.rows.add(new RowData(2, "bbb"));

        SqlFileSelectQuery query = new SqlFileSelectQuery();
        query.setConfig(runtimeConfig);
        query.setSqlFilePath(SqlFiles
                .buildPath(getClass().getName(), getName()));
        query.setCallerClassName("aaa");
        query.setCallerMethodName("bbb");
        query.compile();

        EntityResultListHandler<Emp, Emp_> handler = new EntityResultListHandler<Emp, Emp_>(
                Emp_.class);
        List<Emp> entities = handler.handle(resultSet, query);

        assertEquals(2, entities.size());
        Emp emp = entities.get(0);
        assertEquals(new IntegerDomain(1), emp.id());
        assertEquals(new StringDomain("aaa"), emp.name());
        emp = entities.get(1);
        assertEquals(new IntegerDomain(2), emp.id());
        assertEquals(new StringDomain("bbb"), emp.name());
    }
}
